Update README.md
This repository contains and implemementation of a Split Variational AutoEncoder (SVAE).
In a SVAE the output $\hat{x}$ is computed as a weighted sum $\sigma \odot \hat{x_1} + (1-\sigma) \odot \hat{x_2}$ where $\hat{x_1}, \hat{x_2}$ are two distinct generated
images, and $\sigma$ is a learned compositional map. A Split VAE is trained as a normal VAE: no additional loss is added over the splitted images $\hat{x_1}$ and $\hat{x_2}$. 

Splitting is meant to offer to the network a more flexible way to learn fruitful and independent features: as a result the variable collapse phenomenon is greatly reduced 
and the possibility of exploiting a larger number of latent variables allows to improve the quality and diversity of generated samples.